# api/docs/swagger.yaml
basePath: /aiapp
definitions:
...
  model.ControlAppReq:
    properties:
      action:
        description: Action is one of suspend, resume, reboot, terminate.
        enum:
        - suspend
        - resume
        - reboot
        - terminate
        example: suspend
        type: string
    required:
    - action
    type: object
  model.ControlResult:
    properties:
      action:
        type: string
      affected:
        items:
          type: string
        type: array
      dryRun:
        type: boolean
      infraId:
        type: string
      message:
        type: string
    type: object
  model.DeployAppReq:
    properties:
      appId:
        description: AppID selects the registered application to deploy.
        example: vllm-llama31-8b
        type: string
      imageId:
        description: ImageID pins the node image. Empty asks CB-Tumblebug to select
          one.
        example: ubuntu22.04
        type: string
      infraName:
        description: InfraName becomes the CB-Tumblebug Infra ID and is the idempotency
          key.
        example: vllm-lab-01
        type: string
      nodeCount:
        description: NodeCount is how many serving nodes to provision.
        example: 1
        type: integer
      specId:
        description: SpecID pins the node spec. Empty asks for a recommendation first.
        example: aws+us-west-2+g6.xlarge
        type: string
    required:
    - appId
    - infraName
    type: object
...
  model.DeploymentPlan:
    properties:
      acceleratorSummary:
        description: AcceleratorSummary describes the accelerator the chosen spec
          provides.
        type: string
      appId:
        type: string
      appName:
        type: string
      estimatedCost:
        description: EstimatedCost is the CB-Tumblebug review estimate, when a review
          ran.
        type: string
      imageId:
        type: string
      imageReason:
        description: ImageReason explains an image that was resolved rather than requested.
        type: string
      imageSource:
        description: 'ImageSource records how ImageID was chosen: "requested" or "resolved".'
        type: string
      infraName:
        type: string
      nodeCount:
        type: integer
      nsId:
        type: string
      request:
        description: Request is the dynamic Infra request body, including the install
          commands.
      specId:
        type: string
      specSource:
        description: 'SpecSource records how SpecID was chosen: "requested" or "recommended".'
        type: string
    type: object
  model.DeploymentResult:
    properties:
      dryRun:
        description: DryRun is true when the plan was archived but never sent.
        type: boolean
      infra:
        description: Infra is the provisioned Infra, present only on a real run.
      message:
        description: Message explains the outcome to the caller.
        type: string
      plan:
        $ref: '#/definitions/model.DeploymentPlan'
      review:
        allOf:
        - $ref: '#/definitions/model.ReviewSummary'
        description: Review is the CB-Tumblebug pre-flight verdict, when a review
          ran.
    type: object
...
  model.IntentReq:
    properties:
      intent:
        description: Intent is what the operator wants, in their own words.
        example: Llama 3.1 8B 추론 서버를 GPU 한 장짜리 노드에 올려줘
        type: string
    required:
    - intent
    type: object
  model.IntentResp:
    properties:
      answer:
        description: Answer is what the agent reports back to the operator.
        type: string
      dryRun:
        description: DryRun reflects the guard setting the run executed under.
        type: boolean
      runId:
        type: string
      steps:
        description: Steps is every tool the agent chose, in order, with what the
          guard did.
        items:
          $ref: '#/definitions/model.ToolStep'
        type: array
      usage:
        $ref: '#/definitions/model.TokenUsage'
    type: object
...
  model.ReviewSummary:
    properties:
      creationViable:
        type: boolean
      estimatedCost:
        type: string
      overallMessage:
        type: string
      overallStatus:
        type: string
      recommendations:
        items:
          type: string
        type: array
      totalNodeCount:
        type: integer
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        example: AI application not found
        type: string
    type: object
...
  model.TokenUsage:
    properties:
      inputTokens:
        type: integer
      outputTokens:
        type: integer
    type: object
...
  model.ToolStep:
    properties:
      calledAt:
        type: string
      duration:
        type: string
      error:
        type: string
      executed:
        description: Executed is false when the dry-run guard blocked an infrastructure
          change.
        type: boolean
      grade:
        description: Grade is the tool risk grade that decided whether the call was
          executed.
        type: string
      index:
        type: integer
      input: {}
      output: {}
      tool:
        type: string
    type: object
  tumblebug.InfraInfo:
    properties:
      description:
        type: string
      id:
        type: string
      label:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      node:
        items:
          $ref: '#/definitions/tumblebug.NodeInfo'
        type: array
      status:
        type: string
      statusCount:
        $ref: '#/definitions/tumblebug.StatusCountInfo'
      targetStatus:
        type: string
    type: object
  tumblebug.InfraStatusView:
    properties:
      id:
        type: string
      name:
        type: string
      status:
        type: string
      statusCount:
        $ref: '#/definitions/tumblebug.StatusCountInfo'
      targetAction:
        type: string
      targetStatus:
        type: string
    type: object
  tumblebug.NodeInfo:
    properties:
      id:
        type: string
      imageId:
        type: string
      name:
        type: string
      nodeGroupId:
        type: string
      privateIP:
        type: string
      publicIP:
        type: string
      specId:
        type: string
      sshPort:
        type: integer
      status:
        type: string
    type: object
  tumblebug.StatusCountInfo:
    properties:
      countCreating:
        type: integer
      countFailed:
        type: integer
      countRunning:
        type: integer
      countSuspended:
        type: integer
      countTerminated:
        type: integer
      countTotal:
        type: integer
    type: object
info:
  contact: {}
  description: Registers AI application metadata, turns an operator instruction into
    CB-Tumblebug calls through an LLM, deploys the application onto AI semiconductor
    nodes and archives every step.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AI-MCMP AI Application Automation Agent (Prototype)
  version: 0.1.0
paths:
...
  /ns/{nsId}/deployments:
    get:
      description: List every deployment of the namespace with its node status
      operationId: GetAllDeployments
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment list
          schema:
            items:
              $ref: '#/definitions/tumblebug.InfraInfo'
            type: array
        "500":
          description: Deployment lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List deployments
      tags:
      - '[Deploy] AI Application Deployment'
    post:
      consumes:
      - application/json
      description: Provision nodes and install the AI application on them. The request
        is reviewed first, refused if the name is taken, and only sent to the cloud
        when the dry-run guard is off
      operationId: PostDeployment
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Deployment request
        in: body
        name: deployReq
        required: true
        schema:
          $ref: '#/definitions/model.DeployAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: Deployment planned or provisioned
          schema:
            $ref: '#/definitions/model.DeploymentResult'
        "400":
          description: Malformed request or request not viable
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: AI application not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "409":
          description: Deployment name already in use
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Deployment failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Deploy an AI application
      tags:
      - '[Deploy] AI Application Deployment'
  /ns/{nsId}/deployments/{infraId}:
    delete:
      description: Terminate every node of the deployment and remove its records.
        Shared resources of the namespace are kept
      operationId: DeleteDeployment
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Deployment ID
        in: path
        name: infraId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment deleted or planned
          schema:
            $ref: '#/definitions/model.ControlResult'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Deployment deletion failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete a deployment
      tags:
      - '[Deploy] AI Application Deployment'
  /ns/{nsId}/deployments/{infraId}/control:
    post:
      consumes:
      - application/json
      description: Apply suspend, resume, reboot or terminate to every node. terminate
        stops the nodes but keeps the records, so delete afterwards to stop billing
      operationId: PostDeploymentControl
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Deployment ID
        in: path
        name: infraId
        required: true
        type: string
      - description: Lifecycle action
        in: body
        name: controlReq
        required: true
        schema:
          $ref: '#/definitions/model.ControlAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: Action applied or planned
          schema:
            $ref: '#/definitions/model.ControlResult'
        "400":
          description: Unsupported action
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Deployment control failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Control a deployment
      tags:
      - '[Deploy] AI Application Deployment'
  /ns/{nsId}/deployments/{infraId}/status:
    get:
      description: Report the node status summary, suitable for polling while a deployment
        comes up
      operationId: GetDeploymentStatus
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Deployment ID
        in: path
        name: infraId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deployment status
          schema:
            $ref: '#/definitions/tumblebug.InfraStatusView'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Deployment status lookup failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get deployment status
      tags:
      - '[Deploy] AI Application Deployment'
  /ns/{nsId}/deployments/plan:
    post:
      consumes:
      - application/json
      description: Build the provisioning request for an AI application without contacting
        any cloud. Leave specId empty to select the cheapest spec matching the accelerator
        requirement
      operationId: PostDeploymentPlan
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Deployment request
        in: body
        name: deployReq
        required: true
        schema:
          $ref: '#/definitions/model.DeployAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: Deployment plan
          schema:
            $ref: '#/definitions/model.DeploymentPlan'
        "400":
          description: Malformed or incomplete request
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: AI application not found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Deployment planning failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Plan a deployment
      tags:
      - '[Deploy] AI Application Deployment'
  /ns/{nsId}/intents:
    post:
      consumes:
      - application/json
      description: Hand an operator instruction to the deployment agent. The agent
        resolves the application, recommends an accelerator spec, reviews the request
        and, when the dry-run guard is off, provisions. Every step is archived
      operationId: PostIntent
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Operator instruction
        in: body
        name: intentReq
        required: true
        schema:
          $ref: '#/definitions/model.IntentReq'
      produces:
      - application/json
      responses:
        "200":
          description: Agent run finished
          schema:
            $ref: '#/definitions/model.IntentResp'
        "400":
          description: Malformed or empty instruction
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Agent run failed
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "503":
          description: Planning model is not configured
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Run a natural language deployment intent
      tags:
      - '[Agent] AI Application Automation Agent'
...
swagger: "2.0"