# internal/rest/deployment.go
package rest

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/rs/zerolog/log"

	"github.com/innogrid/ai-agent-proto/internal/model"
)

// malformedBody is the caller-facing text for an unreadable request body.
const malformedBody = "Malformed request body: check JSON syntax"

// allowedActions are the lifecycle actions this API forwards.
//
// Recovery actions (refine, reconcile, abort, withdraw) are excluded: they change
// resource bookkeeping in ways that need a deliberate operator decision.
var allowedActions = map[string]bool{
	"suspend":   true,
	"resume":    true,
	"reboot":    true,
	"terminate": true,
}

// RestPostDeploymentPlan godoc
// @ID PostDeploymentPlan
// @Summary Plan a deployment
// @Description Build the provisioning request for an AI application without contacting any cloud. Leave specId empty to select the cheapest spec matching the accelerator requirement
// @Tags [Deploy] AI Application Deployment
// @Accept json
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Param deployReq body model.DeployAppReq true "Deployment request"
// @Success 200 {object} model.DeploymentPlan "Deployment plan"
// @Failure 400 {object} model.SimpleMsg "Malformed or incomplete request"
// @Failure 404 {object} model.SimpleMsg "AI application not found"
// @Failure 500 {object} model.SimpleMsg "Deployment planning failed"
// @Router /ns/{nsId}/deployments/plan [post]
func (h *Handler) RestPostDeploymentPlan(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")

	req, err := bindDeployReq(c)
	if err != nil {
		return c.JSON(http.StatusBadRequest, model.SimpleMsg{Message: messageFor(err, malformedBody)})
	}

	plan, err := h.deploy.Plan(ctx, nsID, req)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Str("appId", req.AppID).Msg("Failed to plan deployment")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: messageFor(err, "Deployment planning failed")})
	}
	return c.JSON(http.StatusOK, plan)
}

// RestPostDeployment godoc
// @ID PostDeployment
// @Summary Deploy an AI application
// @Description Provision nodes and install the AI application on them. The request is reviewed first, refused if the name is taken, and only sent to the cloud when the dry-run guard is off
// @Tags [Deploy] AI Application Deployment
// @Accept json
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Param deployReq body model.DeployAppReq true "Deployment request"
// @Success 200 {object} model.DeploymentResult "Deployment planned or provisioned"
// @Failure 400 {object} model.SimpleMsg "Malformed request or request not viable"
// @Failure 404 {object} model.SimpleMsg "AI application not found"
// @Failure 409 {object} model.SimpleMsg "Deployment name already in use"
// @Failure 500 {object} model.SimpleMsg "Deployment failed"
// @Router /ns/{nsId}/deployments [post]
func (h *Handler) RestPostDeployment(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")

	req, err := bindDeployReq(c)
	if err != nil {
		return c.JSON(http.StatusBadRequest, model.SimpleMsg{Message: messageFor(err, malformedBody)})
	}

	started := time.Now()
	result, err := h.deploy.Deploy(ctx, nsID, req)
	h.archiveDirect(ctx, nsID, kindDeployment, "deploy_ai_app", started, req, result, h.deploy.DryRun(), err)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Str("appId", req.AppID).
			Str("infraName", req.InfraName).Msg("Failed to deploy AI application")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: messageFor(err, "Deployment failed")})
	}
	return c.JSON(http.StatusOK, result)
}

// RestGetAllDeployments godoc
// @ID GetAllDeployments
// @Summary List deployments
// @Description List every deployment of the namespace with its node status
// @Tags [Deploy] AI Application Deployment
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Success 200 {array} tumblebug.InfraInfo "Deployment list"
// @Failure 500 {object} model.SimpleMsg "Deployment lookup failed"
// @Router /ns/{nsId}/deployments [get]
func (h *Handler) RestGetAllDeployments(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")

	list, err := h.deploy.List(ctx, nsID)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Msg("Failed to list deployments")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: "Deployment lookup failed"})
	}
	return c.JSON(http.StatusOK, list)
}

// RestGetDeploymentStatus godoc
// @ID GetDeploymentStatus
// @Summary Get deployment status
// @Description Report the node status summary, suitable for polling while a deployment comes up
// @Tags [Deploy] AI Application Deployment
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Param infraId path string true "Deployment ID"
// @Success 200 {object} tumblebug.InfraStatusView "Deployment status"
// @Failure 404 {object} model.SimpleMsg "Deployment not found"
// @Failure 500 {object} model.SimpleMsg "Deployment status lookup failed"
// @Router /ns/{nsId}/deployments/{infraId}/status [get]
func (h *Handler) RestGetDeploymentStatus(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")
	infraID := c.Param("infraId")

	status, err := h.deploy.Status(ctx, nsID, infraID)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Str("infraId", infraID).Msg("Failed to get deployment status")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: "Deployment status lookup failed"})
	}
	return c.JSON(http.StatusOK, status)
}

// RestPostDeploymentControl godoc
// @ID PostDeploymentControl
// @Summary Control a deployment
// @Description Apply suspend, resume, reboot or terminate to every node. terminate stops the nodes but keeps the records, so delete afterwards to stop billing
// @Tags [Deploy] AI Application Deployment
// @Accept json
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Param infraId path string true "Deployment ID"
// @Param controlReq body model.ControlAppReq true "Lifecycle action"
// @Success 200 {object} model.ControlResult "Action applied or planned"
// @Failure 400 {object} model.SimpleMsg "Unsupported action"
// @Failure 404 {object} model.SimpleMsg "Deployment not found"
// @Failure 500 {object} model.SimpleMsg "Deployment control failed"
// @Router /ns/{nsId}/deployments/{infraId}/control [post]
func (h *Handler) RestPostDeploymentControl(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")
	infraID := c.Param("infraId")

	var req model.ControlAppReq
	if err := c.Bind(&req); err != nil {
		return c.JSON(http.StatusBadRequest, model.SimpleMsg{Message: malformedBody})
	}
	if !allowedActions[req.Action] {
		return c.JSON(http.StatusBadRequest, model.SimpleMsg{Message: "Action must be suspend, resume, reboot or terminate"})
	}

	started := time.Now()
	result, err := h.deploy.Control(ctx, nsID, infraID, req.Action)
	h.archiveDirect(ctx, nsID, kindControl, "control_deployment", started,
		map[string]any{"infraId": infraID, "action": req.Action}, result, h.deploy.DryRun(), err)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Str("infraId", infraID).Str("action", req.Action).
			Msg("Failed to control deployment")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: "Deployment control failed"})
	}
	return c.JSON(http.StatusOK, result)
}

// RestDeleteDeployment godoc
// @ID DeleteDeployment
// @Summary Delete a deployment
// @Description Terminate every node of the deployment and remove its records. Shared resources of the namespace are kept
// @Tags [Deploy] AI Application Deployment
// @Produce json
// @Param nsId path string true "Namespace ID" default(default)
// @Param infraId path string true "Deployment ID"
// @Success 200 {object} model.ControlResult "Deployment deleted or planned"
// @Failure 404 {object} model.SimpleMsg "Deployment not found"
// @Failure 500 {object} model.SimpleMsg "Deployment deletion failed"
// @Router /ns/{nsId}/deployments/{infraId} [delete]
func (h *Handler) RestDeleteDeployment(c echo.Context) error {
	ctx := c.Request().Context()
	nsID := c.Param("nsId")
	infraID := c.Param("infraId")

	started := time.Now()
	result, err := h.deploy.Delete(ctx, nsID, infraID)
	h.archiveDirect(ctx, nsID, kindControl, "delete_deployment", started,
		map[string]any{"infraId": infraID}, result, h.deploy.DryRun(), err)
	if err != nil {
		log.Error().Err(err).Str("nsId", nsID).Str("infraId", infraID).Msg("Failed to delete deployment")
		return c.JSON(statusCodeFor(err), model.SimpleMsg{Message: "Deployment deletion failed"})
	}
	return c.JSON(http.StatusOK, result)
}

// bindDeployReq reads and validates a deployment request.
func bindDeployReq(c echo.Context) (*model.DeployAppReq, error) {
	var req model.DeployAppReq
	if err := c.Bind(&req); err != nil {
		return nil, fmt.Errorf("failed to bind deployment request: %w", err)
	}
	if err := req.Validate(); err != nil {
		return nil, err
	}
	return &req, nil
}